Moved functions into their folder and started the modularization
